// src/interfaces/relations.ids.interface.ts
export interface RelationsIdsInterface {
  films: number[];
  people: number[];
  planets: number[];
  species: number[];
  starships: number[];
  vehicles: number[];
}

export interface AllSwapiDataInterface {
  // temp any while not wrote interface
  films: any;
  people: any;
  planets: any;
  species: any;
  starships: any;
  vehicles: any;
}